Relocate state functions to state file
